--- a/perfmetrics/scripts/bigquery/constants.py
+++ b/perfmetrics/scripts/bigquery/constants.py
@@ -1,0 +1,22 @@
+# Copyright 2023 Google Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http:#www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Contains constants for bigquery.
+"""
+
+PROJECT_ID = 'gcs-fuse-test-ml'
+DATASET_ID = 'performance_metrics'
+CONFIGURATION_TABLE_ID = 'experiment_configuration'
+FIO_TABLE_ID = 'read_write_fio_metrics'
+VM_TABLE_ID = 'read_write_vm_metrics'
+LS_TABLE_ID = 'list_metrics'