--- a/perfmetrics/scripts/continuous_test/gcp_ubuntu/e2e_tests/e2e-tests-arm64.cfg
+++ b/perfmetrics/scripts/continuous_test/gcp_ubuntu/e2e_tests/e2e-tests-arm64.cfg
@@ -12,4 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+action {
+  define_artifacts {
+    regex: "gcsfuse-failed-integration-test-logs-*"
+    strip_prefix: "github/gcsfuse/perfmetrics/scripts"
+  }
+}
+
 build_file: "gcsfuse/perfmetrics/scripts/continuous_test/gcp_ubuntu/e2e_tests/build.sh"