### Description

### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - Tested locally for bucket mounting, unmounting, listing, stat, read operations
2. Unit tests - tested locally, all passed.
3. Integration tests - ran as part of presubmit.