@gargnitingoogle 2023-10-25
presubmit tests (both perf and integration) passed. Build log [here](https://fusion2.corp.google.com/ci/kokoro/prod:gcsfuse%2Fgcp_ubuntu%2Fpresubmits%2Fperf_tests%2Fpresubmit/activity/e8d9aa7d-4e44-4235-9515-ced990dec2f4/summary) .

Perf results
```
+--------+------------+--------------+------------+--------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
+--------+------------+--------------+------------+--------------+--------------+
| Master |  0.25MiB   | 413.27MiB/s  | 1.07MiB/s  |  31.23MiB/s  |  1.15MiB/s   |
|   PR   |  0.25MiB   | 430.16MiB/s  | 1.12MiB/s  |  33.16MiB/s  |  1.27MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 4751.57MiB/s | 77.96MiB/s | 1042.48MiB/s |  77.12MiB/s  |
|   PR   | 48.828MiB  | 4598.99MiB/s | 80.92MiB/s | 1126.9MiB/s  |  79.6MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4608.75MiB/s | 32.56MiB/s | 528.26MiB/s  |  33.15MiB/s  |
|   PR   | 976.562MiB | 4600.75MiB/s | 28.65MiB/s | 533.01MiB/s  |  31.18MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
+--------+------------+--------------+------------+--------------+--------------+
```